+++ WebServer/startWebServer.py
@@ -7,7 +7,7 @@
 import Domoticz
 
 
-def  startWebServer( self ):
+def  extstartWebServer( self ):
 
     #self.httpPort = '9440'
     self.httpServerConn = Domoticz.Connection(Name="Zigate Server Connection", Transport="TCP/IP", Protocol="HTTP", Port=self.httpPort)
@@ -18,3 +18,48 @@ def  startWebServer( self ):
     #self.httpsServerConn = Domoticz.Connection(Name="Zigate Server Connection", Transport="TCP/IP", Protocol="HTTPS", Port=self.httpsPort)
     #self.httpsServerConn.Listen()
     #self.logging( 'Status', "Web backend for Web User Interface started on port: %s" %self.httpsPort)len(fileContent))+" bytes will be returned")
+
+
+def extonConnect(self, Connection, Status, Description):
+
+    self.logging( 'Debug', "Connection: %s, description: %s" %(Connection, Description))
+    if Status != 0:
+        Domoticz.Error("Failed to connect ("+str(Status)+") to: "+Connection.Address+":"+Connection.Port+" with error: "+Description)
+        return
+
+    # Search for Protocol
+    for item in str(Connection).split(','):
+        if item.find('Protocol') != -1:
+            label, protocol = item.split(':')
+            protocol = protocol.strip().strip("'")
+            self.logging( 'Debug', '%s:>%s' %(label, protocol))
+
+    if protocol == 'HTTP':
+        # http connection
+        if Connection.Name not in self.httpServerConns:
+            self.logging( 'Debug', "New Connection: %s" %(Connection.Name))
+            self.httpServerConns[Connection.Name] = Connection
+    elif protocol == 'HTTPS':
+        # https connection
+        if Connection.Name not in self.httpsServerConns:
+            self.logging( 'Debug', "New Connection: %s" %(Connection.Name))
+            self.httpServerConns[Connection.Name] = Connection
+    else:
+        Domoticz.Error("onConnect - unexpected protocol for connection: %s" %(Connection))
+
+    self.logging( 'Debug', "Number of http  Connections : %s" %len(self.httpServerConns))
+    self.logging( 'Debug', "Number of https Connections : %s" %len(self.httpsServerConns))
+
+
+def extonStop( self ):
+
+    # Make sure that all remaining open connections are closed
+    self.logging( 'Debug', "onStop()")
+
+    # Search for Protocol
+    for connection in self.httpServerConns:
+        self.logging( 'Log', "Closing %s" %connection)
+        self.httpServerConns[Connection.Name].close()
+    for connection in self.httpsServerConns:
+        self.logging( 'Log', "Closing %s" %connection)
+        self.httpServerConns[Connection.Name].close()
